Webavalia to qass conversion (#39)
* Added webavailia to qass conversion mechanism

* Added conversion fix

* Added some design changes

* Fixes for conversion

* Removed comments

* Fixed peer mark table issue

* Added normalized student scores

* Completed int to float conversion for peer marks

+++ db/migrate/20250819114928_create_peer_marks.rb
@@ -5,7 +5,7 @@ class CreatePeerMarks < ActiveRecord::Migration[8.0]
       t.references :group, null: false, foreign_key: true
       t.references :giver, null: false, foreign_key: { to_table: :users }
       t.references :receiver, null: false, foreign_key: { to_table: :users }
-      t.integer :score, null: false
+      t.float :score, null: false
 
       t.timestamps
     end
